Fall 2024/Database/Lab3/requirements.txt
dbis-relational-algebra==1.1.5
docstring-inheritance==2.0.2
gradescope-utils==0.5.0
numpy==1.26.1
pandas==2.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
typeguard==4.1.5
typing_extensions==4.8.0
tzdata==2023.3
